refactor(account-lib): refactor tss folder to easily add new signature schemes
This commit refactors the mpc folder to easily add more signature schemes
by defining dedicated files specific to each implementation.

TICKET:BG-48307
