+++ modules/account-lib/src/mpc/tss/eddsa/index.ts
@@ -1,0 +1,6 @@
+import Eddsa from './eddsa';
+
+// EDDSA Specific types
+export * as EDDSA from './types';
+
+export default Eddsa;
